refactor: Page 하위의 commonStyles를 components 하위의 commonStyles로 이동
