--- src/pages/commonStyles.ts
@@ -1,17 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Title = styled.h1`
-  font-size: 24px;
-  font-weight: 700;
-`;
-
-export const SubTitle = styled.h2`
-  font-size: 16px;
-  font-weight: 700;
-`;
-
-export const Description = styled.p`
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 18px;
-`;
